+++ NEWS.md
@@ -11,6 +11,10 @@
   created bugs!!!
   
 - New vignette *Timeseries Covariates with NSGEV*. DRAFT version.
+
+- A warning is cast when the `design` argument of `TVGEV` is a call
+  using variables not in `names(data)`. These variables may or may not
+  be found in the parent environment.
   
 ## Bug fixes
  
@@ -30,6 +34,10 @@
   
 - The Hessian of the negalive log-likelihood is now exact and no longer 
   computed with `optimHessian`.
+  
+- The `.Rnw` vignettes, such as the technical vignette `CompDet.Rnw`
+  should no longer be built. Only the pdf will be shipped with the 
+  package tarball.
 
 # New in version 0.2.0